moved the util functions to same dir/names as libSBML
