+++ src/bindings/swig/liblx.h
@@ -63,7 +63,7 @@
 #include <liblx/xml/LibLXError.h>
 
 #include <liblx/common/operationReturnValues.h>
-#include <liblx/common/util.h>
+#include <liblx/util/util.h>
 
 #include <liblx/xml/XMLAttributes.h>
 #include <liblx/xml/XMLBuffer.h>
