+++ src/liblx/xml/XMLTriple.cpp
@@ -37,7 +37,7 @@
 #include <liblx/xml/XMLTriple.h>
 #include <liblx/xml/XMLAttributes.h>
 #include <liblx/xml/XMLConstructorException.h>
-#include <liblx/common/util.h>
+#include <liblx/util/util.h>
 
 /** @cond doxygenIgnored */
 using namespace std;
