+++ src/liblx/xml/test/TestXMLAttributesC.c
@@ -48,13 +48,12 @@
 #include <XMLAttributes.h>
 #include <XMLTriple.h>
 #include <XMLErrorLog.h>
-//#include <util/memory.h>
-//#include <util/util.h>
+#include <util/memory.h>
+#include <util/util.h>
 #include <check.h>
 
 #include <liblx/common/extern.h>
 
-#include <liblx/common/util.h>
 
 #if defined(__cplusplus)
 LIBLX_CPP_NAMESPACE_USE
