+++ src/liblx/xml/test/TestXMLToken.c
@@ -39,6 +39,7 @@
 #include <liblx/xml/XMLTriple.h>
 
 #include <liblx/common/operationReturnValues.h>
+#include <liblx/util/memory.h>
 
 #include <check.h>
