+++ src/liblx/xml/test/TestXMLToken_newSetters.c
@@ -48,6 +48,7 @@
 
 
 #include <liblx/common/extern.h>
+#include <liblx/util/memory.h>
 
 #if defined(__cplusplus)
 LIBLX_CPP_NAMESPACE_USE
